requirements.txt
annotated-types==0.7.0
anyio==4.9.0
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
arrow==1.3.0
asttokens==3.0.0
async-lru==2.0.5
attrs==23.2.0
Babel==2.10.3
backoff==2.2.1
bcc==0.29.1
bcrypt==5.0.0
beautifulsoup4==4.13.4
bleach==6.2.0
blinker==1.9.0
Brlapi==0.8.5
build==1.3.0
cachetools==6.2.2
certifi==2023.11.17
cffi==1.17.1
chardet==5.2.0
chromadb==1.3.5
click==8.1.6
cloud-init==25.2
colorama==0.4.6
coloredlogs==15.0.1
comm==0.2.2
command-not-found==0.3
configobj==5.0.8
cryptography==41.0.7
cupshelpers==1.0
dbus-python==1.3.2
debugpy==1.8.14
decorator==5.2.1
defer==1.0.6
defusedxml==0.7.1
distro==1.9.0
distro-info==1.7+build1
durationpy==0.10
executing==2.2.0
fastjsonschema==2.21.1
filelock==3.20.0
Flask==3.1.2
flatbuffers==25.9.23
fqdn==1.5.1
fsspec==2025.10.0
fuse-python==1.0.7
git-filter-repo==2.38.0
google-api-core==2.28.1
google-api-python-client==2.187.0
google-auth==2.41.1
google-auth-httplib2==0.2.1
google-auth-oauthlib==1.2.3
googleapis-common-protos==1.72.0
grpcio==1.76.0
gyp==0.1
h11==0.16.0
hf-xet==1.2.0
httpcore==1.0.9
httplib2==0.20.4
httptools==0.7.1
httpx==0.28.1
huggingface_hub==1.1.5
humanfriendly==10.0
idna==3.6
importlib_metadata==8.7.0
importlib_resources==6.5.2
ipykernel==6.29.5
ipython==9.3.0
ipython_pygments_lexers==1.1.1
ipywidgets==8.1.7
isoduration==20.11.0
itsdangerous==2.2.0
jedi==0.19.2
Jinja2==3.1.2
json5==0.12.0
jsonpatch==1.32
jsonpointer==2.0
jsonschema==4.24.0
jsonschema-specifications==2025.4.1
jupyter==1.1.1
jupyter-console==6.6.3
jupyter-events==0.12.0
jupyter-lsp==2.2.5
jupyter_client==8.6.3
jupyter_core==5.8.1
jupyter_server==2.16.0
jupyter_server_terminals==0.5.3
jupyterlab==4.4.3
jupyterlab_pygments==0.3.0
jupyterlab_server==2.27.3
jupyterlab_widgets==3.0.15
kubernetes==34.1.0
language-selector==0.1
launchpadlib==1.11.0
lazr.restfulclient==0.14.6
lazr.uri==1.0.6
louis==3.29.0
markdown-it-py==3.0.0
MarkupSafe==2.1.5
matplotlib-inline==0.1.7
mdurl==0.1.2
mistune==3.1.3
mmh3==5.2.0
mpmath==1.3.0
nbclient==0.10.2
nbconvert==7.16.6
nbformat==5.10.4
nest-asyncio==1.6.0
netaddr==0.8.0
netifaces==0.11.0
notebook==7.4.3
notebook_shim==0.2.4
numpy==2.3.5
oauthlib==3.2.2
olefile==0.46
ollama==0.6.1
onnxruntime==1.23.2
opentelemetry-api==1.38.0
opentelemetry-exporter-otlp-proto-common==1.38.0
opentelemetry-exporter-otlp-proto-grpc==1.38.0
opentelemetry-proto==1.38.0
opentelemetry-sdk==1.38.0
opentelemetry-semantic-conventions==0.59b0
orjson==3.11.4
overrides==7.7.0
packaging==25.0
pandocfilters==1.5.1
parso==0.8.4
pexpect==4.9.0
pillow==10.2.0
platformdirs==4.3.8
posthog==5.4.0
prometheus_client==0.22.0
prompt_toolkit==3.0.51
proto-plus==1.26.1
protobuf==6.33.1
psutil==7.0.0
ptyprocess==0.7.0
pure_eval==0.2.3
pyasn1==0.6.1
pyasn1_modules==0.4.2
pybase64==1.4.2
pycairo==1.25.1
pycparser==2.22
pycups==2.0.1
pydantic==2.12.4
pydantic_core==2.41.5
Pygments==2.17.2
PyGObject==3.48.2
PyJWT==2.7.0
pylibacl==0.7.0
pyparsing==3.1.1
PyPika==0.48.9
pyproject_hooks==1.2.0
PyQt5==5.15.10
PyQt5-sip==12.13.0
pyrsistent==0.20.0
pyserial==3.5
python-apt==2.7.7+ubuntu5
python-dateutil==2.8.2
python-debian==0.1.49+ubuntu2
python-dotenv==1.0.1
python-json-logger==3.3.0
python-linux-procfs==0.6.3
pytz==2024.1
pyudev==0.24.0
pyxattr==0.8.1
pyxdg==0.28
PyYAML==6.0.1
pyzmq==26.4.0
referencing==0.36.2
requests==2.31.0
requests-oauthlib==2.0.0
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rich==13.7.1
rpds-py==0.25.1
rsa==4.9.1
screen-resolution-extra==0.0.0
Send2Trash==1.8.3
setuptools==68.1.2
shellingham==1.5.4
six==1.16.0
sniffio==1.3.1
soupsieve==2.7
stack-data==0.6.3
sympy==1.14.0
systemd-python==235
tenacity==9.1.2
terminado==0.18.1
tinycss2==1.4.0
tokenizers==0.22.1
tornado==6.4
tqdm==4.67.1
traitlets==5.14.3
typer==0.20.0
typer-slim==0.20.0
types-python-dateutil==2.9.0.20250516
typing-inspection==0.4.2
typing_extensions==4.15.0
ubuntu-drivers-common==0.0.0
ubuntu-pro-client==8001
ufw==0.36.2
unattended-upgrades==0.1
uri-template==1.3.0
uritemplate==4.2.0
urllib3==2.0.7
uvicorn==0.38.0
uvloop==0.22.1
wadllib==1.3.6
waitress==3.0.2
watchfiles==1.1.1
wcwidth==0.2.13
webcolors==24.11.1
webencodings==0.5.1
websocket-client==1.8.0
websockets==15.0.1
Werkzeug==3.1.3
wheel==0.42.0
widgetsnbextension==4.0.14
xdg==5
xkit==0.0.0
zipp==3.23.0
